Create setup debug function for local env
