+++ api/SweeperSmackdown/.gitignore
@@ -262,3 +262,6 @@ paket-files/
 # Python Tools for Visual Studio (PTVS)
 __pycache__/
 *.pyc
+
+# Additional rules
+!Functions/Http/[Dd]ebug/
