+++ api/SweeperSmackdown/Extensions/CosmosClientExtensions.cs
@@ -2,6 +2,7 @@ using Microsoft.Azure.Cosmos;
 using Microsoft.Azure.Cosmos.Linq;
 using SweeperSmackdown.Assets;
 using SweeperSmackdown.Models;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -54,6 +55,12 @@ public static class CosmosClientExtensions
             .ReadAllAsync();
     }
 
+    /// <summary>
+    /// Change the host of a lobby to a different active player.
+    /// </summary>
+    /// <param name="cosmosClient">The CosmosClient</param>
+    /// <param name="userId">The ID of the host that is leaving</param>
+    /// <returns>An enumerable of all lobbies given a new host</returns>
     public static async Task<IEnumerable<Lobby>> ChangeHostAsync(this CosmosClient cosmosClient, string userId)
     {
         var lobbies = await cosmosClient
@@ -67,9 +74,9 @@ public static class CosmosClientExtensions
             lobbies.Select(lobby => Task.Run(async () => {
                 var players = await cosmosClient.GetAllPlayersInLobbyAsync(lobby.Id);
 
-                if (players.Any())
+                if (players.Any(p => p.Active))
                 {
-                    var hostId = players.First().Id;
+                    var hostId = players.First(p => p.Active).Id;
 
                     await cosmosClient.GetLobbyContainer().PatchItemAsync<Lobby>(lobby.Id, new(lobby.Id), new[]
                     {
@@ -84,5 +91,45 @@ public static class CosmosClientExtensions
             })));
 
         return updatedLobbies.Where(l => l is not null) as IEnumerable<Lobby>;
+    }
+
+    /// <summary>
+    /// Regenerate database containers. For debug purposes only.
+    /// </summary>
+    /// <param name="cosmosClient">The CosmosClient</param>
+    /// <returns>A task which resolves once all containers have been regenerated</returns>
+    public static async Task RegenerateContainers(this CosmosClient cosmosClient)
+    {
+        var database = cosmosClient.GetSmackdownDatabase();
+
+        // Setup lobby container
+        try
+        {
+            await cosmosClient.GetLobbyContainer().DeleteContainerAsync();
+        }
+        catch (Exception)
+        {
+        }
+
+        await database.CreateContainerAsync(new()
+        {
+            Id = DatabaseConstants.LOBBY_CONTAINER_NAME,
+            PartitionKeyPath = "/id"
+        });
+
+        // Setup player container
+        try
+        {
+            await cosmosClient.GetPlayerContainer().DeleteContainerAsync();
+        }
+        catch (Exception)
+        {
+        }
+
+        await database.CreateContainerAsync(new()
+        {
+            Id = DatabaseConstants.PLAYER_CONTAINER_NAME,
+            PartitionKeyPath = "/lobbyId"
+        });
     }
 }
